[INLONG-4301]fix some comments
